refactor(rolldown_watcher): distinguish debounced and non-debounced watchers
